# Conflicts:
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/common/WorkTimezoneCommonSetSetMemento.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktime/predset/PredetemineTimeSettingRepository.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/worktime/KshmtFlexStampReflectPK.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/worktime/predset/JpaPredetemineTimeSetRepository.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/worktime/worktimeset/JpaWorkTimeSettingRepository.java

+++ nts.uk/uk.com/com.web/nts.uk.com.web/src/main/webapp/view/cps/001/c/cps001.c.vm.ts
@@ -15,9 +15,9 @@ module cps001.c.vm {
     export class ViewModel {
         listEmployee: KnockoutObservableArray<IEmployee> = ko.observableArray([]);
         currentEmployee: KnockoutObservable<Employee> = ko.observable(new Employee());
-        
-        enaBtnRes : KnockoutObservable<boolean> = ko.observable(true);
-        enaBtnDel : KnockoutObservable<boolean> = ko.observable(true);
+
+        enaBtnRes: KnockoutObservable<boolean> = ko.observable(true);
+        enaBtnDel: KnockoutObservable<boolean> = ko.observable(true);
 
         constructor() {
             let self = this,
@@ -26,6 +26,8 @@ module cps001.c.vm {
 
             emp.id.subscribe(x => {
                 if (x) {
+                    self.enableControl();
+
                     let iem: IEmployee = _.find(self.listEmployee(), e => e.id == x);
 
                     service.getDetail(x).done((data: IEmployee) => {
@@ -36,8 +38,11 @@ module cps001.c.vm {
 
                             emp.reason(data.reason || '');
                             emp.dateDelete(data.dateDelete || undefined);
+                            $('#code').focus();
                         }
                     });
+                } else {
+                    self.newMode();
                 }
             });
 
@@ -51,11 +56,10 @@ module cps001.c.vm {
                 emp = self.currentEmployee();
 
             emps.removeAll();
-
             service.getData().done((data: Array<IEmployee>) => {
                 if (data && data.length) {
                     emps(data);
-
+                    $('#code').focus();
                     if (!sid) {
                         emp.id(data[0].id);
                     } else {
@@ -73,6 +77,7 @@ module cps001.c.vm {
                     self.newMode();
                 }
                 dfd.resolve();
+            }).fail(() => {
             });
             return dfd.promise();
         }
@@ -82,6 +87,10 @@ module cps001.c.vm {
                 emp: IEmployee = ko.toJS(self.currentEmployee),
                 listItem: Array<IEmployee> = ko.toJS(self.listEmployee);
 
+            if (nts.uk.ui.errors.hasError()) {
+                return;
+            }
+
             confirm({ messageId: "Msg_528" }).ifYes(() => {
                 let itemListLength = self.listEmployee().length,
                     indexItemDelete = _.findIndex(ko.toJS(self.listEmployee), function(item: any) { return item.id == emp.id; }),
@@ -90,21 +99,24 @@ module cps001.c.vm {
                         code: emp.code,
                         name: emp.name
                     };
-
+                block();
                 service.restoreData(objToRestore).done(() => {
                     if (itemListLength === 1) {
                         self.start();
                     } else if (itemListLength - 1 === indexItemDelete) {
-                        self.start(listItem[indexItemDelete - 1].id).done(() => {
-                        });
+                        self.start(listItem[indexItemDelete - 1].id);
                     } else if (itemListLength - 1 > indexItemDelete) {
-                        self.start(listItem[indexItemDelete + 1].id).done(() => {
-                        });
+                        self.start(listItem[indexItemDelete + 1].id);
                     }
+
+                    unblock();
+
+                }).fail((mes) => {
+                    unblock();
                 });
 
             }).ifCancel(() => {
-
+                unblock();
             });
         }
 
@@ -113,8 +125,14 @@ module cps001.c.vm {
                 emp: IEmployee = ko.toJS(self.currentEmployee()),
                 listItem: Array<IEmployee> = ko.toJS(self.listEmployee());
 
+            if (nts.uk.ui.errors.hasError()) {
+                return;
+            }
+
+
             confirm({ messageId: "Msg_18" }).ifYes(() => {
                 let sid = emp.id;
+                block();
                 service.removedata(sid).done(() => {
                     showDialog.info({ messageId: "Msg_464" }).then(function() {
                         let itemListLength = self.listEmployee().length,
@@ -122,18 +140,19 @@ module cps001.c.vm {
                         if (itemListLength === 1) {
                             self.start();
                         } else if (itemListLength - 1 === indexItemDelete) {
-                            self.start(listItem[indexItemDelete - 1].id).done(() => {
-                            });
+                            self.start(listItem[indexItemDelete - 1].id);
                         } else if (itemListLength - 1 > indexItemDelete) {
-                            self.start(listItem[indexItemDelete + 1].id).done(() => {
-                            });
+                            self.start(listItem[indexItemDelete + 1].id);
                         }
+                        unblock();
                     });
 
 
+                }).fail((mes) => {
+                    unblock();
                 });
             }).ifCancel(() => {
-
+                unblock();
             });
         }
 
@@ -142,10 +161,10 @@ module cps001.c.vm {
         }
 
         newMode() {
-           let self = this,
+            let self = this,
                 emps = self.listEmployee(),
                 emp = self.currentEmployee();
-            
+
             emp.enableCode(false);
             emp.enableName(false);
             emp.code('');
@@ -155,6 +174,19 @@ module cps001.c.vm {
             self.enaBtnRes(false);
             self.enaBtnDel(false);
         }
+
+        enableControl() {
+            let self = this,
+                emps = self.listEmployee(),
+                emp = self.currentEmployee();
+
+            emp.enableCode(true);
+            emp.enableName(true);
+            self.enaBtnRes(true);
+            self.enaBtnDel(true);
+        }
+
+
     }
 
     interface IEmployee {
